requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
httpx==0.26.0
psycopg2-binary
python-multipart
email-validator
bcrypt==4.0.1